refactor: endpoint에 따른 RequestMapping 구현
